no message
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@496 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/client.cpp
@@ -353,7 +353,7 @@ void gets2c()           // get updates from the server
         case ENET_EVENT_TYPE_RECEIVE:
             if(disconnecting) 
 				conoutf("attempting to disconnect...");
-            else localservertoclient(event.channelID, event.packet->data, event.packet->dataLength);
+            else localservertoclient(event.channelID, event.packet->data, (int)event.packet->dataLength);
             enet_packet_destroy(event.packet);
             break;
